[ 008 ] Variable Expansion
Tokens that starts with an '$' sign will now be treated as a variable. And no metter it is a single token ("/home/yakub") or in the middle of the token ("abc/home/yakub,abc") it should be evaluated. What's more, there're few new builtins: pwd, env, unset. export builtin is being made