[ 016 ] rcfiles: beginning
This is an erly implementation of rcfiles: program reads (or creates) an rcfile from the /home/yakub directory and proceeds to read it line-by-line. For now it only prints the content of the file, but in the future I want to ad a complete execution of each line using the framework from the normals shell implementation.