[ 017 ] rcfiles: execution
For now, each line of the '~/.shrc' file will be executed as a standard command. The exception is the comment line, started with the '#' sign, which will be ignored. Also, the file descriptor leaks are now fixed and the only possible leaks of this program are by the utilities and by the readline library