[ 026 ] Statusline improvements
Statusline prompting was significantly improved. The current version allows me to extend the parsing process to other escape sequences (for example: I can add color text as a string code). What's more, using regex I can retrieve the name of the operating system, which then is used by the desktop format, replicating the behaviour of other shells. The last problem that is there since forever is to fix the problem with broken environment when this shell is set as default login shell.